pyproject.toml
[project]
name = "redis-cache-service-py"
version = "0.1.0"
description = "Async Redis-based caching service with event-driven invalidation and concurrency demo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "typer>=0.19.2",
]

